winkernel: add proxier_util.go
